+++ R/xrd.get.AllPeaks.R
@@ -16,8 +16,6 @@
 #' @return vector with peak positions
 #'
 #' @seealso [xrd.find.Peak()]
-#' @importFrom dplyr filter "%>%"
-#' @importFrom rlang .data
 #'
 #' @examples
 #'
@@ -32,10 +30,9 @@ xrd.get.AllPeaks <- function(data,
                              deltaTheta = 5,
                              Range = c(0,90),
                              verbose = FALSE) {
+  dataXRD <- check_dataXRD(data)
   # XRD data
-  d <- data.frame(TwoTheta = data$theta,
-                  Intensity = data$I) %>%
-    filter(.data$TwoTheta >= Range[1] & .data$TwoTheta <= Range[2])
+  d <- xrd_filter(dataXRD, Range[1], Range[2])
 
   step.size = 2 # search peaks with this step size
   theta.min = min(d$TwoTheta) + step.size/2
@@ -53,9 +50,9 @@ xrd.get.AllPeaks <- function(data,
   }
 
   for(ang in peakPos.list) {
-    n1 <- d %>% filter(.data$TwoTheta >= (ang-deltaTheta) & .data$TwoTheta<= (ang+deltaTheta))
+    n1 <- xrd_filter(d, ang-deltaTheta, ang+deltaTheta)
 
-    p1 = xrd.find.Peak(n1$TwoTheta, n1$Intensity,
+    p1 = xrd.find.Peak(n1$TwoTheta, n1$I,
                        Try.Sigma = Try.Sigma,
                        peakPos = ang,
                        verbose = verbose)