+++ R/xrd.get.PeakStats.R
@@ -1,17 +1,17 @@
 #' Fits a Gaussian Function to XRD data
 #'
 #' @description
-#' usually run after a peak is
-#' identified through its location
+#' Given an XRD peak, a Gaussian fit at the provided location \code{peakPos}
+#' a Gaussian fit is made to find the exact location, width, and background.
 #'
 #' @author Thomas Gredig
 #'
-#' @param TwoTheta 2q angle
-#' @param Intensity XRD intensity
-#' @param peakPos specific 2q angle of peak position
+#' @param TwoTheta vector of 2Theta angles
+#' @param Intensity vector of XRD intensities
+#' @param peakPos specific 2Theta angle of peak position
 #' @param Try.Sigma vector with peak widths used to start fitting
 #' @param verbose logical, if \code{TRUE} output additional information
-#'
+#'x
 #' @return background, amplitude, position, width + 4 std. errors
 #' @examples
 #' filename = xrd.getSampleFiles(fileExt='asc')
@@ -21,10 +21,12 @@
 #' @importFrom stats nls sd predict
 #' @importFrom graphics lines
 #' @export
-xrd.get.PeakStats <- function(TwoTheta, Intensity, peakPos,
+xrd.get.PeakStats <- function(TwoTheta,
+                              Intensity = NULL,
+                              peakPos,
                               Try.Sigma = c(0.2,0.1,0.05,0.3),
                               verbose = FALSE) {
-  d = data.frame(TwoTheta, I = Intensity)
+  d <- check_dataXRD(TwoTheta, Intensity)
   for(pw in Try.Sigma) {
 
     deltaPeak = max(0.2, 4*pw)
@@ -32,9 +34,9 @@ xrd.get.PeakStats <- function(TwoTheta, Intensity, peakPos,
     n1 = subset(d, TwoTheta > (peakPos-deltaPeak) &
                   TwoTheta < (peakPos+deltaPeak))
     if (verbose) plot(n1)
-    if (verbose) print(paste("Fit from 2q =",
+    if (verbose) cat("Fit from 2 theta =",
                              min(n1$TwoTheta), "to",
-                             max(n1$TwoTheta)))
+                             max(n1$TwoTheta),"\n")
     if(nrow(n1)>5) {
       q = xrd.peakEstimate(n1$TwoTheta, n1$I)
       background = ceiling(q$b0)
@@ -43,10 +45,10 @@ xrd.get.PeakStats <- function(TwoTheta, Intensity, peakPos,
       peak.width = pw
 
       if (verbose) {
-        print(paste("background=",background,
+        cat("background=",background,
                     " A=",A1,
                     " t0=",p1,
-                    " sigma=",peak.width))
+                    " sigma=",peak.width,"\n")
       }
       try(fit <-
             nls(data = n1,